submissions/submission_hateclipper/requirements.txt
torch
torchvision
pytesseract
opencv-python-headless
scipy
scikit-image
ninja
numpy
easyocr
tqdm
pyyaml
transformers
pytorch_lightning
pillow